packages/engine/pyproject.toml
[tool.poetry]
name = "amethyst-engine"
version = "0.1.0"
description = "Amethyst execution engine"
authors = ["Arko Bhattacharya <arko@fask.ai>"]
readme = "README.md"
packages = [{include = "amethyst_engine", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
a2a-sdk = ">=0.2.14,<0.3.0"
uvicorn = ">=0.35.0,<0.36.0"
fastapi = ">=0.115.0,<0.116.0"
grpcio = ">=1.73.1,<2.0.0"
protobuf = ">=6.31.1,<7.0.0"
googleapis-common-protos = ">=1.70.0,<2.0.0"
mcp = ">=1.1.0,<2.0.0"
openai = ">=2.6.0,<3.0.0"
python-dotenv = ">=1.0.0,<2.0.0"
pipedream = ">=1.0.10,<2.0.0"
pydantic = ">=2.12.3,<3.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["F", "E", "I"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["amethyst_engine"]
